requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-multipart
firebase-admin
pydantic-settings
python-dotenv